LPS-17978
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@83031 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portlet/journal/service/impl/JournalArticleLocalServiceImpl.java
@@ -642,9 +642,11 @@ public class JournalArticleLocalServiceImpl
 
 		// Workflow
 
-		workflowInstanceLinkLocalService.deleteWorkflowInstanceLink(
-			article.getCompanyId(), article.getGroupId(),
-			JournalArticle.class.getName(), article.getResourcePrimKey());
+		if (!article.isDraft()) {
+			workflowInstanceLinkLocalService.deleteWorkflowInstanceLink(
+				article.getCompanyId(), article.getGroupId(),
+				JournalArticle.class.getName(), article.getResourcePrimKey());
+		}
 
 		int articlesCount = journalArticlePersistence.countByG_A(
 			article.getGroupId(), article.getArticleId());
